+++ test.sh
@@ -1,0 +1,8 @@
+#!/usr/bin/env bash
+
+set -e
+
+rm -rf .box-venv
+mkdir .box-venv
+
+sandbox -v $(pwd):/pwd -v ~/.mitmproxy:/home/sprrw/.mitmproxy -v $(pwd)/.box-venv:/pwd/.venv -- bash -c 'cd /pwd; exec bash'